Predicate1 implemented with macro annotation
