+++ macros/src/main/scala/com/wheaties/predicate/Macros.scala
@@ -1,17 +1,11 @@
 package com.wheaties.predicate
 
-import annotation.{compileTimeOnly, StaticAnnotation, Annotation}
+import annotation.{compileTimeOnly, StaticAnnotation}
 import scala.reflect.macros.whitebox
 import scala.reflect.macros._
 import scala.language.experimental.macros
 
 object Macros {
-//	def hello: Unit = macro impl
-//
-//  def impl(c: whitebox.Context): c.Expr[Unit] = {
-//    import c.universe._
-//    c.Expr(q"""println("Hello World")""")
-//  }
 
   @compileTimeOnly("GeneratePredicate must be called during compilation")
   class GeneratePredicate extends StaticAnnotation {
@@ -28,8 +22,46 @@ object Macros {
         println(s"bazinga ${classDecl.name}")
         c.Expr(
           q"""
-            trait ${classDecl.name} {
-              def print() = println("hello annotations")
+            trait ${classDecl.name}[T1] extends Function1[T1, Boolean]{
+              self =>
+
+              def or[TT1 <: T1](that: Function1[TT1, Boolean]) = new Predicate1[TT1]{
+                def apply(arg1: TT1) = self(arg1) || that(arg1)
+              }
+              def and[TT1 <: T1](that: Function1[TT1, Boolean]) = new Predicate1[TT1]{
+                def apply(arg1: TT1) = self(arg1) && that(arg1)
+              }
+              def xor[TT1 <: T1](that: Function1[TT1, Boolean]) = new Predicate1[TT1]{
+                def apply(arg1: TT1) = if(self(arg1)) !that(arg1) else that(arg1)
+              }
+              def nor[TT1 <: T1](that: Function1[TT1, Boolean]) = new Predicate1[TT1]{
+                def apply(arg1: TT1) = !(self(arg1) || that(arg1))
+              }
+              def nand[TT1 <: T1](that: Function1[TT1, Boolean]) = new Predicate1[TT1]{
+                def apply(arg1: TT1) = !(self(arg1) && that(arg1))
+              }
+              def nxor[TT1 <: T1](that: Function1[TT1, Boolean]) = new Predicate1[TT1]{
+                def apply(arg1: TT1) = if(self(arg1)) that(arg1) else !that(arg1)
+              }
+              def not = new Predicate1[T1]{
+                def apply(arg1: T1) = !self(arg1)
+              }
+              override def toString() = "<predicate1>"
+
+            }
+
+            object ${classDecl.name.toTermName} {
+              object Always extends ${classDecl.name}[Any]{
+                def apply(arg1: Any) = true
+              }
+
+              val always = Always
+
+              object Never extends ${classDecl.name}[Any]{
+                def apply(arg1: Any) = false
+              }
+
+              val never = Never
             }
            """)
       }
@@ -40,31 +72,6 @@ object Macros {
       }
     }
 
-  }
-
-
-  object helloMacro {
-    def impl(c: Context)(annottees: c.Expr[Any]*): c.Expr[Any] = {
-      import c.universe._
-      import Flag._
-      val result = {
-        annottees.map(_.tree).toList match {
-          case q"object $name extends ..$parents { ..$body }" :: Nil =>
-            q"""
-            object $name extends ..$parents {
-              def hello: ${typeOf[String]} = "hello"
-              ..$body
-            }
-          """
-        }
-      }
-      c.Expr[Any](result)
-    }
-  }
-
-  @compileTimeOnly("some msg")
-  class hello extends StaticAnnotation {
-    def macroTransform(annottees: Any*) = macro helloMacro.impl
   }
 }
