+++ predicates/src/test/scala/com/wheaties/predicate/PredicateSpec.scala
@@ -63,17 +63,20 @@ class PredicateSpec extends WordSpecLike
     }
 
     "have implicit `not` method" in new SpecExamples {
-      val filtered = sampleList.filter(Predicate1.not.not(Modulo(2, 0)))
+//      val filtered = sampleList.filter(Predicate1.not.not(Modulo(2, 0)))
+      val filtered = sampleList.filter(Modulo(2, 0).not)
       filtered should equal(List(3, 5))
     }
 
     "have accompanying Always" in new SpecExamples {
-      val filtered = sampleList.filter(Modulo(2, 0) or Always1)
+//      val filtered = sampleList.filter(Modulo(2, 0) or Always1)
+      val filtered = sampleList.filter(Modulo(2, 0) or Predicate1.always)
       filtered should equal(sampleList)
     }
 
     "have accompanying Never" in new SpecExamples {
-      val filtered = sampleList.filter(Modulo(2, 0) and Never1)
+//      val filtered = sampleList.filter(Modulo(2, 0) and Never1)
+      val filtered = sampleList.filter(Modulo(2, 0) and Predicate1.never)
       filtered should equal(List.empty)
     }
   }
